+++ rust/polars-hs-ffi/src/lib.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::missing_safety_doc)]
 #![allow(clippy::not_unsafe_ptr_arg_deref)]
 
+pub mod arrow;
 pub mod bytes;
 pub mod dataframe;
 pub mod error;